Adds some details to the readme.
Adds World Data that deals with world generation. In this case, specifically the ores. Since the Geysers are still in testing, they are omitted to mitigate people getting erroneous results later if they depend on that information.

Changes the Issue Report [Bug] to indicate versions not listed are no longer supported. This is typically caused by lack of time, or environments have changed drastically. `1.20.1` is an example where the ENTIRE code base for the build system and structure was flipped upside down for better clarity, but this means that backporting is not really a targeted goal.

Closes #74 As the requested data is part of this commit (in theory)